Don't store our font scale factor as a percentage
While we want to show the font scale factor as a percentage in the UI, internally, we should represent it as a CGFloat, which is more consistent with the NSFont APIs and makes interacting with them easier.

+++ Frameworks/OakTextView/src/OakTextView.h
@@ -35,7 +35,7 @@ PUBLIC @interface OakTextView : OakView
 @property (nonatomic) theme_ptr                             theme;
 @property (nonatomic) NSCursor*                             ibeamCursor;
 @property (nonatomic) NSFont*                               font;
-@property (nonatomic) NSInteger                             fontScaleFactor;
+@property (nonatomic) CGFloat                               fontScaleFactor;
 @property (nonatomic) BOOL                                  antiAlias;
 @property (nonatomic) OTVFontSmoothing                      fontSmoothing;
 @property (nonatomic) size_t                                tabSize;
